fix: refactored signer selector and filter components to use a base component
