+++ src/components/SignerSelector/SignerSelectorTs.ts
@@ -18,13 +18,11 @@ import { Component, Prop, Vue } from 'vue-property-decorator';
 // @ts-ignore
 import FormRow from '@/components/FormRow/FormRow.vue';
 import { Signer } from '@/store/Account';
-
-export class SignerItem {
-    constructor(public parent: boolean, public signer: Signer, public level: number) {}
-}
+// @ts-ignore
+import SignerBaseFilter from '@/components/SignerBaseFilter/SignerBaseFilter.vue';
 
 @Component({
-    components: { FormRow },
+    components: { FormRow, SignerBaseFilter },
 })
 export class SignerSelectorTs extends Vue {
     /**
@@ -72,30 +70,9 @@ export class SignerSelectorTs extends Vue {
     get chosenSigner(): string {
         return this.value;
     }
-
-    /**
-     * @returns {SignerItem[]} multisig parent signers
-     */
-    get multisigSigners(): SignerItem[] {
-        return this.getParentSignerItems(this.rootSigner, 0);
-    }
-
-    /**
-     * Starts with the given signer, traverse multisig(parent) signers recursively and returns flat signer tree
-     * @param {Signer} signer
-     * @param {number} level
-     * @returns {SignerItem[]} Signer tree as a list
-     */
-    public getParentSignerItems(signer: Signer, level: number): SignerItem[] {
-        if (!signer?.parentSigners) {
-            return [];
-        }
-        const signerItems: SignerItem[] = [];
-        for (const parentSigner of signer.parentSigners) {
-            signerItems.push(new SignerItem(parentSigner.parentSigners?.length > 0, parentSigner, level));
-            signerItems.push(...this.getParentSignerItems(parentSigner, level + 1));
-        }
-        return signerItems;
-    }
     /// end-region computed properties getter/setter
+
+    public onSignerSelectorChange(newSigner: string) {
+        this.chosenSigner = newSigner;
+    }
 }
